reworked the code - deleted unnecessary functions and all helpers to a separete file
