+++ matrix/src/s21_calc_complements.c
@@ -21,9 +21,7 @@ int s21_calc_complements(matrix_t *A, matrix_t *result) {
       } else {
           matrix_t temp = {0};
           double temp_res = 0.0;
-          if (s21_create_matrix(A->rows, A->columns, &temp)) {
-            return_code = ERROR_ENUM;
-          } else {
+          s21_create_matrix(A->rows, A->columns, &temp);
             for (int row_i = 0; row_i < A->rows; ++row_i) {
               for (int col_i = 0; col_i < A->columns; ++col_i) {
                 get_cofactor(A->matrix, temp.matrix, row_i, col_i, A->rows);
@@ -32,14 +30,9 @@ int s21_calc_complements(matrix_t *A, matrix_t *result) {
                 result->matrix[row_i][col_i] = temp_res;
               }
             }
-          s21_remove_matrix(&temp);
-        }
+          s21_remove_matrix(&temp);                                
       }
     }
   }
   return return_code;
-}
-
-void get_algebraic_complement(int row, int col, double* res) {
-  *res *= pow(-1.0, row + col + 2);
 }
